requirements.txt
aiofiles==24.1.0
aiogram==3.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.2
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
frozenlist==1.8.0
greenlet==3.3.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.2
pandas==3.0.1
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.47
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
w3lib==2.4.0
Werkzeug==3.1.6
xgboost==3.2.0
yarl==1.22.0
